+++ inlong-audit/audit-service/src/main/java/org/apache/inlong/audit/selector/api/Selector.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.audit.elector.api;
+package org.apache.inlong.audit.selector.api;
 
 /**
  * Selector
