+++ inlong-audit/audit-service/src/main/java/org/apache/inlong/audit/selector/api/SelectorConfig.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.audit.elector.api;
+package org.apache.inlong.audit.selector.api;
 
 import lombok.Data;
 import org.apache.commons.lang3.StringUtils;
