+++ inlong-audit/audit-service/src/main/java/org/apache/inlong/audit/selector/api/SelectorFactory.java
@@ -15,21 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.audit.elector;
+package org.apache.inlong.audit.selector.api;
 
-import org.apache.inlong.audit.elector.api.SelectorChangeListener;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.inlong.audit.selector.impl.SelectorImpl;
 
 /**
- * Elector change listener impl
+ * Selector factory
  */
-public class ElectorChangeListenerImpl implements SelectorChangeListener {
+public class SelectorFactory {
 
-    private static final Logger logger = LoggerFactory.getLogger(ElectorChangeListenerImpl.class);
-
-    public void leaderChanged(boolean currentNodeIsLeader) {
-        logger.info("LeaderChanged {}:", currentNodeIsLeader);
+    public static Selector getNewElector(SelectorConfig electorConfig) {
+        return new SelectorImpl(electorConfig);
     }
 }
